Add the refactored modules import path to the strategy module
